+++ src/main/java/com/arka/catalog_service/infrastructure/adapters/repository/ProductosJpaRepository.java
@@ -7,17 +7,17 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductosJpaRepository extends JpaRepository<ProductosEntity,Integer> {
+public interface ProductosJpaRepository extends JpaRepository<ProductosEntity,Long> {
 
 
     Productos create(Productos producto);
-    Optional<ProductosEntity> findById(Integer id);
+    Optional<ProductosEntity> findById(Long id);
     Optional<ProductosEntity> findByNombre(String nombre);
     List<ProductosEntity> findAllByMarca(String marca);
-    List<ProductosEntity> findAllByCategoria(Integer categoria);
+    List<ProductosEntity> findAllByCategoria(Long categoria);
     List<ProductosEntity>findAll();
     boolean existsByNombre(String nombre);
-    boolean existsById(Integer id);
-    boolean existsByNombreAndCategoria(String nombre, Integer categoria);
+    boolean existsById(Long id);
+    boolean existsByNombreAndCategoria(String nombre, Long categoria);
 
 }
